requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
python-jose[cryptography]
python-dotenv
openai
pillow
requests
requests-oauthlib